refactor: move outline specific drag helpers to outline folder
