+++ src/course-outline/drag-helper/SortableItem.jsx
