+++ src/course-outline/drag-helper/SortableItem.scss
