+++ src/course-outline/drag-helper/utils.js
