+++ src/generic/styles.scss
@@ -11,6 +11,5 @@
 @import "./tag-count/TagCount";
 @import "./modal-dropzone/ModalDropzone";
 @import "./configure-modal/ConfigureModal";
-@import "./drag-helper/SortableItem";
 @import "./block-type-utils";
 @import "./modal-iframe"
